+++ merkle/README.md
@@ -1,0 +1,5 @@
+# NOTE
+
+This implementation is forked from cbergoon's implementation. I claim no authorship over this code apart from some minor modifications.
+
+An implementation of a Merkle Tree written in Go. A Merkle Tree is a hash tree that provides an efficient way to verify the contents of a set data are present and untampered with.
