+++ merkle/hash.go
